Delay DTLS handshake to force and test specific scenario
This is a tiny hack artifically delaying certain steps in the DTLS
handshake to arrive at the commissioning sequence described in one of
the GitHub issues.

To test, run the following from the Mbed TLS base directory:

> ./programs/test/udp_proxy_wrapper.sh pack=10 -- server_addr=localhost
  server_port=4433 dtls=1 event=1

for the server, and then

> ./programs/ssl/ssl_client2 server_port=4433 server_addr=localhost dtls=1
  event=1

for the client, and inspect what happens with Wireshark.

Without the splitting WANT_READ in WANT_READ and CONTINUE_PROCESSING, this the
handshake will deadlock; incorporating the change, it'll succeed.