Streamline reuse of old messages
The usual behaviour or the record parsing functions is to discard the last record and to process a new one. In few
places this is not the desired behaviour, for example when checking for optional messages in a handshake or when
receiving application data while waiting for renegotiation confirmation; in these cases instead, the current package has
to be held back to be further processed by the dedicated routines.

This commit unifies and simplifies the code handling these situations.