Prototype handshake test in ssl test suite
This commit adds a test to the SSL test suite performing a PSK-based SSL/TLS handshake. Client and server take turns
within the same process and exchange data through primitive read- and write-buffers. The intend of this is to be able to
give a framework to test corner cases or anomalous behavior easily.