Remove bad-signature test files
This commit removes the original test files and Readme related to checking that non-reduced length encodings in RSA PKCS v1.5 signatures are
not tolerated. The contents of these files had been duplicated in the RSA test suites, which however couldn't be
documented until the introduction of comments in test files. Now that these are available, the origin of the data and the
details of the respective test cases are commented upon in the RSA test suite.