Restructure handling of client's ServerHello state
This commit exemplifies a potential restructuring in the SSL/TLS logic
layer in the example of the client's ServerHello state handling.

In general, the intention is to split monolithic functions like
`ssl_parse_server_hello` into multiple functions performing the various
stages of the message parsing, processing, and state transitioning; in
addition to increasing readability, the purpose of this is to pave the
way for

a) a more flexible unit testing infrastructure for the SSL/TLS module,
b) forthcoming reworkings of the message processing stack.

In the example at hand, we split `ssl_parse_server_hello` as follows:

* Main entry: `ssl_process_server_hello`
  The top-level function called when the client is in ServerHello state.
  It is responsible for orchestrating the helper functions explained next.

- Preprocessing: `ssl_process_server_hello_coordinate`
  This function attempts to fetch a new record, checks the type of the
  incoming message and does some bookkeeping on the number of unexpected
  records (no parsing).

- Parsing:    `ssl_process_server_hello_parse`
           OR `ssl_process_hello_verify_parse`

  The actual parsing functions.

- Postprocessing: inlined so far
  * Checks the renegotiation state and usage of
    secure renegotiation indication.
  * Derives keys if a session is being resumed.
  * Updates state machine, flights and checksums.

Move handshake failure check to postprocessing of serverhello