Compute outgoing MAC in temporary buffer for Encrypt-then-MAC
A previous commit changed the record encryption function
`ssl_encrypt_buf` to compute the MAC in a temporary buffer
and copying the relevant part of it (which is strictly smaller
if the truncated HMAC extension is used) to the outgoing message
buffer. However, the change was only made in case MAC-then-Encrypt
was enabled, but not in case of Encrypt-then-MAC.

For the sake of uniformity, this commit changes `ssl_encrypt_buf`
to compute the MAC in a temporary buffer in the Encrypt-then-MAC
case, too.

In addition to uniformity, this also avoids `mbedtls_ssl_encrypt_buf()`
temporarily writing beyond the bounds of current outgoing record in
case the truncated HMAC extension is used (note, though, that this
does not come at the danger of a heap buffer overflow, as the size
computation for the outgoing record buffer accounts for enough space
to hold a non-truncated MAC in the buffer.