!TEST ONLY! Avoid record buffer overflow in `ssl_decrypt_buf`
The MAC checking code in the record decryption routine `ssl_decrypt_buf`
may overflow the record content buffer, which is caught by the MPS Layer 2
random tests; because of additional padding, though, it cannot overflow
the heap buffer that contains the record -- the overflow is therefore
not critical to be fixed immediately, but should be dealt with at some
point.

This TEST-ONLY commit removed the respective code-path.

!!! WARNING !!!
This is a temporary, test-only commit, and MUST NOT be put to production,
because the commented code-path is a security measure against cache
timing attacks.