Restructure handling of incoming CertificateRequest message
This commit restructures the handling of the incoming
CertificateRequest message on the client.

Specifically, it renames the client-side monolithic function
ssl_parse_certificate_request to ssl_process_certificate_request
and splits the latter into the following sub-functions:

* Coordination

  In CertificateRequest state, we don't know whether to expect
  a CertificateRequest message or a ServerHelloDone. Also, some
  ciphersuites don't allow CertificateRequest message to be
  sent at all.

  The coordination step decides whether a CertificateRequest message
  is expected by checking if a handshake message has arrived and
  peeking at its handshake content type byte. If the byte matches
  the CertificateRequest handshake type, we expect a valid
  CertificateRequest message. If not, we silently skip this step
  (progressing to ServerHelloDone).

* Parsing & Validation

  Parse and validate the CertificateRequestmessage
  and remember that client authentication is needed
  by setting the `client_auth` field in the SSL context.

* Postprocessing: ssl_certificate_request_postprocess

  Update the handshake state machine.

The test script `tests/ssl-opt.sh` needed to be slightly adapted:
Previously, the message "skip parse certificate request" was printed
only if the ciphersuite didn't allow certificate requests, but
it wasn't if a certificate request was allowed but not used.
The present commit prints "skip parse certificate request" whenever
the parsing of the certificate request is omitted, regardless of the
reason, and this breaks the "SNI: client auth override: optional -> none",
"SNI: DTLS, client auth override" optional -> none"
and "Authentication: client badcert, server none" tests in
`tests/ssl-opt.sh` which grep for it. To correct this, an
additional level-3 debug message "ciphersuite doesn't allow certificate
request" is added to the certificate request handling (in addition to the
existing "skip parse certificate request" one) which the respective
tests now check for.