Restructure In-CRT: Move further checks for client-CRT skipping
This commit moves the code responsible for skipping client
authentication on the server in the case where
- The RSA_PSK ciphersuite is used
- The authentication mode (potentially overwritten by SNI)
  is VERIFY_NONE
from ssl_parse_certificate() to ssl_read_certificate_coordinate().

The RSA_PSK ciphersuite is not among those handled in the previous
commit, because it uses server authentication but omits client
authentication (at least our implementation does -- the RFC for
RSA_PSK is not clear on that).