Restructure In-CRT: Move SSLv3 Client-No-CRT code
This commit moves the code-path responsible for handling
the case where SSLv3 is used and the client doesn't present
a certificate to authenticate itself.

In contrast to TLS 1.0 and higher, where an empty certificate
chain is sent, SSLv3 uses a dedicated alert.

The restructured code attempts to unify the treatment of the
client not presenting a certificate by reflecting it in an
unset peer-CRT structure. The peer-CRT structure is setup
within the CRT parsing routine, so the possibilities for the
peer-CRT structure to be left unset are
- the parsing routine is skipped,
- the parsing routine is called but finds an empty chain.
The former happens if a NoCert alert is received for SSLv3,
and the latter happens for TLS 1.0 and higher.

The peer-CRT structure is inspected by the CRT validation routine
ssl_read_certificate_validate(), which in particular deals with
the case where it is unset, rejecting it for mandatory authentication
and accepting it for optional or unset authentication.

Finally, the test case in tests/ssl-opt.sh triggering the situation
where an SSLv3 client can't present a certificate and sends a dedicated
No-Cert alert instead is enhanced by a check that the server indeed
receives and reads that alert.