Restructure In-CRT: Move certificate parsing code
This commit moves the certificate parsing code from
ssl_parse_certificate() to ssl_read_certificate_parse().

A notable change is the treatment of empty certificates:
The modified code accepts an empty certificate chain,
reflecting it in an unset peer-CRT structure. It is up
to the final validation routine ssl_read_certificate_validate()
to check whether the missing certificate can be tolerated or not.