Restructure outgoing CrtReq: Writing 1/2, copy code
This commit moves the code-path responsible for writing the
outgoing CertificateRequest message from ssl_write_certificate_request()
to the new writing function ssl_certificate_request_write().

IMPORTANT: Bounds checks were previously omitted, and could be so
in any sane configuration, because the CertificateRequest message
was always written to the beginning of the output buffer.
Now that the writing function is standalone, and that concretely
the CertificateRequest message might be piggy-backed in the same
record as other handshake messages, we must add bounds checks.
This will be done in the next commit.