Writer: Allow pass-through of commits to be configurable
When a handshake message is written the total length of which is
not known in advance, the handshake header cannot yet be written
at the time the handshake record is opened, but instead must
be postponed until the handshake message has been written and
the size is known. Since committing invalidates the output buffers
obtained from the writer, this means that the handshake writing
process must not use any commits.

On the one hand, this is not a big restriction, because handshake
messages whose length is not known in advance cannot be paused
anyway, and committing at the level of the extended writer is mainly
a means to mark pre-emptable steps during the writing process.

On the other hand, for a writing function taking a large output buffer
and returning the actual amount of written data only later, there must
be a way to inform the writer about how much of the output buffer
obtained from mbedtls_writer_get_ext() was actually used, i.e.
to restrict the output buffer in retrospect.
Since the output buffer returned by the writer for the restricted
length might be different from the restricted output buffer originally
returned, such restriction can only be implemented if either
(a) it goes alongside the commitment of not writing to the restricted
    output buffer anymore, or
(b) the writer implementation remembers the list of
    restricted output buffers.
Variant (b) would increase the complexity of the writer implementation
significantly, hence we go with alternative (a), i.e. restricting a
previously fetched output buffer while at the same time committing it.
This functionality has previously been implemented as
mbedtls_writer_commit_partial().

However, at the time the user performs this partial commit, the
handshake header hasn't been completed yet, so we must not
passthrough the partial commit to the underlying reader immediately.
But if we don't do that, we must not let the user continue using
the writer anymore, as extended and underlying writer got out
of sync.

This commit suggests and implements the following solution
to this problem:
(a) When the length of the handshake message is not known in
    advance, the extended writer is configured in such a way
    that it remembers, but does not pass through, commits
    and partial commits. Moreover, once a single partial commit
    is performed, it cannot be used anymore.
    This is restrictive, but enables the most basic use case
    in which a user fetches an output buffer as large as possible,
    calls the writing function on it, and partially commits what the
    writing function has written.
    The number of ordinary commits is unlimited -- note that
    these are useless as long as pausing is not used, which can't
    be done in case the handshake length is not known in advance
    anyway. However, allowing commits even if the handshake length
    is not known might ease the transition to writing functions
    for which we know the handshake length in advance and make
    use of pausing.

    This mode of operation is enabled by passing the additional
    argument MBEDTLS_WRITER_EXT_HOLD when calling
    mbedtls_writer_init_ext().

    The information on how much data was committed, and how
    much was left uncommitted, is now returned when the
    underlying writer is reclaimed via mbedtls_writer_reclaim().

(b) If the total length of the handshake message is known,
    the handshake header can be written in advance. However,
    it should still not yet be committed because it might be
    that the user finds that there's not enough space to write
    the handshake message, and abort the write instead.

    There are two options in this case:
    (1) Again don't forward any commits from the extended
        to the underlying writer, and allow only a single
        partial commit.
    (2) Pass through all commits, including partial ones,
        thereby committing the handshake header as part of
        the first commit made to the extended writer.

    Both approaches are possible, but currently we follow
    approach (2) as gives full flexibility regarding partial
    commits to the writing functions.

    This full pass-through mode iis enabled by passing the additional
    argument MBEDTL_WRITER_EXT_PASS when calling
    mbedtls_writer_init_ext().