UDP proxy: Use out buffer structure even if !TIMING_C
Previously, the out buffer structure for packing outgoing
datagrams was only used when MBEDTLS_TIMING_C was enabled
and the pack options was specified.

This commit unconditionally puts the output buffer structure
to use, regardless of the setting of packing option.

The reasons for this are (1) uniformity, and (2) that when
datagram packing is used and the UDP proxy knows how to split
incoming datagrams into records, we might want to mimick
the initial record packing again when forwarding the
potentially modified records; for example, we might want
to simulate reordering two records within a single datagram,
to ensure that the messaging stack doesn't make any assumptions
on ordering just because records arrive in the same datagram.