TLS-1.2 PRF: Safe a copy of key within generator structure
The previous implementation of the TLS 1.2 PRF kept the key slot
identifier within the generator struct, checking and reading the
key in each key expansion step.

For the use of the TLS-1.2 PRF with the TLS-1.2 PSK-to-MasterSecret
expansion, we need to call the TLS-1.2 PRF with the padded PSK, which
a priori is not available in any key slot. Therefore, the present
implementation of TLS-1.2 PRF could only be used if a temporary
key slot for the padded PSK was created.

This commit paves the way for a different approach: Instead of the
key slot identifier, the TLS-1.2 PRF generator structure now maintains
a copy of the key material. This copy is created from the raw key
that is passed to `psa_generator_tls12_prf_setup()` from
`psa_key_derivation()`.

This change will allow the TLS-1.2 PSK-to-MS derivation to call
`psa_generator_tls12_prf_setup()` with the padded PSK without
creating a separate key slot for it.

As a side effect, this change makes the TLS-1.2 PRF agnostic to
destruction of the key slot during key expansion.